Fix hang when swapping positions on chunk boundary
Npc_actor::step would cache the actor's chunk before trying to swap positions
to resolve a blocked state. If this swap caused the chunk to change,
then Npc_actor::step would call Map_chunk::remove() on the wrong chunk,
corrupting the state of the new current chunk of the actor.