Use a screenView for "SDL_Window"
SDL always use full screen size for window in iOS,
but in our case we need to honor the window size request from Exult.
And we need to do the scaling in SDL side, therefore Exult should
always create the minimal display surface, and let the SDL side
do the scaling.