fix #9 walking stopped while dpad is down
The root cause is that the game play needs continuous key down events
to drive the character moving forward. In real keyboard, when you hold
down a key, key events will be generated repeatedly. Our emulated dpad
should behave in the same way so that the character can move smoothly.