Add shortcut bar gump at the top of screen
We have been using the native input dialog for text input.
The keyboard button gump is not actually used at the moment.
So we just reuse the keyboard button gump for shortcut bar gump.
If we need to bring back keyboard button in future,
we can simply add it to the shortcut bar.

The disk button item is used for showing menu, double click will open
the load/save dialog.

toggle combat&map teleport is also implemented.