fix #21 shortcut bar disappear
Commit  5293c1f  tried to fix the crash issue, but the shortcut bar is
missing (apparently my device must have gone seriously wrong that I
didn’t notice it!).  This is an attempt to correct it.

The solution: still we don’t want to add shortcut bar to gump manager’s
gump list because it will try to close all gumps when loading a saved
games. Following the foot steps of gkeyboard(CE version), we draw the
shortcut bar in Gump_manager::paint(),  and handle events in
Gump_manager::handle_modal_gump_event().