Allowing specification of (experimental, buggy, long disabled) usecode debugger in configure with
	--with-usecode-debugger=no|console|yes

'console' enables a console debugger that *might* work, but is untested;
'yes' enables the ES debugger, which only has the Exult side (no ES UI)
and hence will not work.

Fixed all errors and warnings on the debugger's code, and re-enabled the
bits that were disabled with #if 0.

TODO: make it all work and test it.