More improvements to data sources and their usage
Renamed old IBufferDataSource to IBufferDataView, which is a non-owning
*view* of data.

New IBufferDataSource is now an owning source of data, and takes ownership
of the buffer it uses.

Renamed old OBufferDataSource to OBufferDataSpan, which is a non-owning
writable span of data (blame C++20).

New OBufferDataSource is now an owning sink of data, and takes ownership
of the buffer it uses.

Modified code to account for this.